Make more API-ish
