+++ app/controllers/api/v1/customers_controller.rb
@@ -1,0 +1,8 @@
+module Api
+  module V1
+    class CustomersController < ApplicationController
+      def index
+      end
+    end
+  end
+end
