--- app/controllers/customers_controller.rb
@@ -1,4 +1,0 @@
-class CustomersController < ApplicationController
-  def index
-  end
-end
